move buttons to parent component
